Finished Data Analysis
Developed a script to extract wind data in low-motion times for the sail and gondola, but there isn't much there, and the data that is there indicates that there is no guarantee that the pitot tube is pointing into the wind. In the interest of my thesis schedule, I'm going to close out this branch and start developing the sail and gondola sim model, with the understanding that wind data will probably be simulated entirely or will be based on a one-second dataset gathered in Utah in 2010-2011.